# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Bolo",
        "description": "Projeto de cadastro de bolo e disparo de email",
        "contact": {
            "email": "fabriciopps19@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API V1"
        }
    ],
    "paths": {
        "/api/v1/cake": {
            "get": {
                "tags": [
                    "Cake"
                ],
                "summary": "Get list of Cake",
                "description": "Returns list of Cake",
                "operationId": "getCakesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CakeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CakeResource": {
                "title": "CakeResource",
                "description": "Project resource",
                "properties": {
                    "": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CakeResource"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Cake",
            "description": "API Endpoints"
        }
    ]
}